loot-alert-api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
redis==5.0.7
requests==2.32.3
exponent-server-sdk==2.1.0
stripe==10.5.0
python-dotenv==1.0.1
bcrypt==4.1.3
gunicorn==22.0.0
APScheduler==3.10.4
feedparser==6.0.11
PyJWT==2.8.0
